py-feos exports of Datasets, moved records from props to sets.
