+++ crates/feos-core/src/ad/properties/liquid_density.rs
@@ -1,6 +1,5 @@
 use crate::DensityInitialization::Liquid;
 use crate::ad::Gradient;
-use crate::ad::dataset::DatasetRecord;
 use crate::ad::{ParametersAD, vectorize, vectorize_ad};
 use crate::density_iteration::density_iteration;
 use crate::{FeosResult, ReferenceSystem, Residual, State};
@@ -8,30 +7,6 @@ use nalgebra::U1;
 use ndarray::{Array1, Array2, ArrayView2};
 use num_dual::DualStruct;
 use quantity::{Density, KELVIN, KILO, METER, MOL, Moles, PASCAL, Pressure, Temperature};
-use serde::{Deserialize, Serialize};
-
-#[derive(Deserialize, Serialize)]
-pub struct LiquidDensityRecord {
-    pub temperature_k: f64,
-    pub pressure_pa: f64,
-    pub liquid_density_kmol_m3: f64,
-}
-
-impl DatasetRecord for LiquidDensityRecord {
-    const N_INPUTS: usize = 2;
-
-    fn input(&self, column: usize) -> f64 {
-        match column {
-            0 => self.temperature_k,
-            1 => self.pressure_pa,
-            _ => unreachable!("invalid liquid density input column"),
-        }
-    }
-
-    fn target(&self) -> f64 {
-        self.liquid_density_kmol_m3
-    }
-}
 
 pub fn liquid_density_ad<E: Residual<U1, Gradient<P>>, const P: usize>(
     eos: &E,
